requirements.txt
streamlit==1.46.1
python-dotenv==1.1.1
web3==7.14.0
openai==2.3.0
requests==2.32.5
networkx==3.5
pyvis==0.3.2
matplotlib==3.10.7
numpy==2.3.1
scikit-learn==1.7.2
plotly==6.3.1
weasyprint==66.0
